+++ controllers/front/Ajax.php
@@ -11,6 +11,7 @@
  */
 
 use Invertus\dpdBaltics\Config\Config;
+use Invertus\dpdBaltics\Exception\DpdCarrierException;
 use Invertus\dpdBaltics\Provider\CurrentCountryProvider;
 use Invertus\dpdBaltics\Repository\ParcelShopRepository;
 use Invertus\dpdBaltics\Repository\ProductRepository;
@@ -49,8 +50,7 @@ class DpdBalticsAjaxModuleFrontController extends ModuleFrontController
             $isAjaxParameter &&
             $isTokenValid &&
             $isParent &&
-            $isAjaxRequest &&
-            $isCustomer;
+            $isAjaxRequest;
 
         if (!$isAccessGranted) {
             $error = $this->module->l('Unauthorized access', self::FILENAME);
@@ -103,9 +103,55 @@ class DpdBalticsAjaxModuleFrontController extends ModuleFrontController
                 $pudoId = Tools::getValue('id_pudo');
                 $this->savePudoPickupPoint($pudoId, $countryCode);
                 break;
+            case 'validateOrderCustom':
+                $response = false;
+                /** @var \Invertus\dpdBaltics\Validate\Carrier\PudoValidate $pudoValidate */
+                $pudoValidator = $this->module->getModuleContainer('invertus.dpdbaltics.validate.carrier.pudo_validate');
+                /** @var  \Invertus\dpdBaltics\Validate\Phone\PhoneNumberValidator $phoneNumberValidator */
+                $phoneNumberValidator = $this->module->getModuleContainer('invertus.dpdbaltics.validate.phone.phone_number_validator');
+                try {
+                    $carrier = new Carrier($carrierId);
+                    $cartId = Context::getContext()->cart->id;
+
+                    $phoneNumberValidator->isPhoneAddedInOrder($cartId);
+                    $pudoValidator->isPudoSelected($cartId, $carrier->id_reference);
+                } catch (DpdCarrierException $exception) {
+                    $this->setErrorMessage($exception);
+                    $this->ajaxDie(json_encode(
+                        [
+                            'status' => false,
+                            'template' => $this->getMessageTemplate('danger'),
+                        ]
+                    ));
+                }
+                $this->ajaxDie(json_encode(['status' => true]));
+                break;
             case 'updateStreetSelect':
                 $city = Tools::getValue('city');
                 $this->updateStreetSelect($countryCode, $city);
+                break;
+            case 'saveSelectedPhoneNumber':
+                $cartId = Context::getContext()->cart->id;
+                /** @var  \Invertus\dpdBaltics\Service\CarrierPhoneService $phoneService */
+                $phoneService = $this->module->getModuleContainer('invertus.dpdbaltics.service.carrier_phone_service');
+                /** @var  \Invertus\dpdBaltics\Validate\Phone\PhoneNumberValidator $phoneNumberValidator */
+                $phoneNumberValidator = $this->module->getModuleContainer('invertus.dpdbaltics.validate.phone.phone_number_validator');
+                $response = false;
+                try {
+                    $prefix = Tools::getValue('phone_area');
+                    $phone = Tools::getValue('phone_number');
+                    $phoneNumberValidator->isPhoneValid($prefix, $phone);
+                    $response = $phoneService->saveCarrierPhone($cartId, $phone, $prefix);
+                } catch (DpdCarrierException $exception) {
+                    $this->setErrorMessage($exception);
+                    $this->ajaxDie(json_encode(
+                        [
+                            'status' => false,
+                            'template' => $this->getMessageTemplate('danger'),
+                        ]
+                    ));
+                }
+                $this->ajaxDie(json_encode(['status' => true]));
                 break;
             case 'updateParcelBlock':
                 $street = Tools::getValue('street');
@@ -324,5 +370,34 @@ class DpdBalticsAjaxModuleFrontController extends ModuleFrontController
             'selectedPudoId' => $selectedPudo->getParcelShopId(),
             'coordinates' => $coordinates
         ]));
+    }
+
+    /**
+     * @param Exception | DpdCarrierException $exception
+     */
+    private function setErrorMessage($exception)
+    {
+        switch ($exception->getMessage()) {
+            case Config::ERROR_COULD_NOT_SAVE_PHONE_NUMBER:
+                $this->messages[] = $this->module->l('Could not save phone number');
+                break;
+            case Config::ERROR_BAD_PHONE_NUMBER_PREFIX:
+                $this->messages[] = $this->module->l('Phone number prefix is empty');
+                break;
+            case Config::ERROR_PHONE_EMPTY:
+                $this->messages[] = $this->module->l('Phone number is empty');
+                break;
+            case Config::ERROR_PHONE_HAS_INVALID_CHARACTERS:
+                $this->messages[] = $this->module->l('Phone number contains invalid characters');
+                break;
+            case Config::ERROR_PHONE_HAS_INVALID_LENGTH:
+                $this->messages[] = $this->module->l('Phone number length is invalid');
+                break;
+            case Config::ERROR_INVALID_PUDO_TERMINAL:
+                $this->messages[] = $this->module->l('Pudo point is missing, please select valid terminal point');
+                break;
+            default:
+                $this->messages[] = $exception->getMessage();
+        }
     }
 }