+++ src/Entity/DPDZone.php
@@ -60,7 +60,8 @@ class DPDZone extends ObjectModel
      */
     public static function checkAddressInZones(Address $address, array $zones)
     {
-        $country = new Country($address->id_country);
+        $idCountry = $address->id_country ?: (int)Configuration::get('PS_COUNTRY_DEFAULT');
+        $country = new Country($idCountry);
 
         /** @var DPDBaltics $module */
         $module = Module::getInstanceByName('dpdbaltics');